+++ src/FluxFlow.Components.FileSystem/Contracts/DirectoryEntryType.cs
@@ -1,0 +1,7 @@
+namespace FluxFlow.Components.FileSystem.Contracts;
+
+public enum DirectoryEntryType
+{
+    File,
+    Directory
+}